video_summarizer/requirements.txt
gradio
opencv-python
ffmpeg-python
torch
numpy
Pillow
git+https://github.com/openai/CLIP.git
transformers
accelerate